NLP: moved Query class to separate file
