+++ src/nlp/src/api/query.py
@@ -1,0 +1,77 @@
+from typing import Optional
+from pydantic.dataclasses import dataclass
+
+
+@dataclass
+class BaseAttributes:
+    min: Optional[int]
+    max: Optional[int]
+
+    def __init__(self):
+        self.min = 0
+        self.max = 0
+
+
+@dataclass
+class GradiantAttributes(BaseAttributes):
+    length: Optional[int]
+
+    def __init__(self):
+        super().__init__()
+        self.length = 0
+
+
+@dataclass
+class CurveAttributes(BaseAttributes):
+    count: Optional[int]
+
+    def __init__(self):
+        super().__init__()
+        self.count = 0
+
+
+@dataclass
+class Curves(CurveAttributes):
+    left: Optional[CurveAttributes]
+    right: Optional[CurveAttributes]
+
+    def __init__(self):
+        super().__init__()
+        self.left = CurveAttributes()
+        self.right = CurveAttributes()
+
+
+@dataclass
+class RouteAttributes:
+    location_start: Optional[str]
+    location_end: Optional[str]
+    height: Optional[BaseAttributes]
+    length: Optional[BaseAttributes]
+    gradiant: Optional[GradiantAttributes]
+    curves: Optional[Curves]
+    charging_stations: bool
+    toll_road_avoidance: bool
+
+    def __init__(self):
+        self.height = BaseAttributes()
+        self.length = BaseAttributes()
+        self.gradiant = GradiantAttributes()
+        self.curves = Curves()
+        self.charging_stations = False
+        self.toll_road_avoidance = False
+        self.location_start = ""
+        self.location_end = ""
+
+
+@dataclass
+class Query:
+    location: Optional[str]
+    max_distance: Optional[int]
+    query_object: str
+    route_attributes: Optional[RouteAttributes]
+
+    def __init__(self) -> None:
+        self.location = ""
+        self.max_distance = 0
+        self.query_object = ""
+        self.route_attributes = RouteAttributes()
