+++ src/nlp/src/api/string_interpreter.py
@@ -5,14 +5,13 @@ LOGGER = logging.getLogger("[STRING INTERPRETER]")
 import os
 import pathlib
 import sys
-from typing import Optional
 
 import spacy
-from pydantic.dataclasses import dataclass
 
 from .helper_service import convert_number_to_meter, check_similarity, check_similarity_in_list, \
     get_keyword_from_synonyms
 from .utils import get_entity_synonyms, get_alias_synonyms
+from .query import Query
 
 # get os specific file separator
 SEP = os.path.sep
@@ -423,81 +422,6 @@ def check_negation(prev_token: str) -> bool:
         if check_similarity(synonym, prev_token, threshold=0.8):
             return True
     return False
-
-
-@dataclass
-class BaseAttributes:
-    min: Optional[int]
-    max: Optional[int]
-
-    def __init__(self):
-        self.min = 0
-        self.max = 0
-
-
-@dataclass
-class GradiantAttributes(BaseAttributes):
-    length: Optional[int]
-
-    def __init__(self):
-        super().__init__()
-        self.length = 0
-
-
-@dataclass
-class CurveAttributes(BaseAttributes):
-    count: Optional[int]
-
-    def __init__(self):
-        super().__init__()
-        self.count = 0
-
-
-@dataclass
-class Curves(CurveAttributes):
-    left: Optional[CurveAttributes]
-    right: Optional[CurveAttributes]
-
-    def __init__(self):
-        super().__init__()
-        self.left = CurveAttributes()
-        self.right = CurveAttributes()
-
-
-@dataclass
-class RouteAttributes:
-    location_start: Optional[str]
-    location_end: Optional[str]
-    height: Optional[BaseAttributes]
-    length: Optional[BaseAttributes]
-    gradiant: Optional[GradiantAttributes]
-    curves: Optional[Curves]
-    charging_stations: bool
-    toll_road_avoidance: bool
-
-    def __init__(self):
-        self.height = BaseAttributes()
-        self.length = BaseAttributes()
-        self.gradiant = GradiantAttributes()
-        self.curves = Curves()
-        self.charging_stations = False
-        self.toll_road_avoidance = False
-        self.location_start = ""
-        self.location_end = ""
-
-
-@dataclass
-class Query:
-    location: Optional[str]
-    max_distance: Optional[int]
-    query_object: str
-    route_attributes: Optional[RouteAttributes]
-
-    def __init__(self) -> None:
-        self.location = ""
-        self.max_distance = 0
-        self.query_object = ""
-        self.route_attributes = RouteAttributes()
 
 
 class TokenKeywords:
